+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.23.0] - 2025-07-15
+
+### Changed
+
+- Added `--filter-include-content-type-regex` option.
+- Added `--filter-exclude-content-type-regex` option.
+
 ## [1.22.1] - 2025-07-14
 
 ### Changed